refactor: server-base too
