+++ packages/server/core/src/base/middlewares/customServer/base.ts
@@ -7,7 +7,7 @@ import {
 } from '@modern-js/types';
 import { getCookie } from 'hono/cookie';
 import { HonoContext, HonoRequest, Logger } from '../../../core/server';
-import { getHost } from '../request';
+import { getHost } from '../../libs/request';
 
 export function createBaseHookContext(
   c: HonoContext,
