+++ packages/server/core/src/base/middlewares/index.ts
@@ -1,4 +1,4 @@
 export * from './faviconFallback';
-export * from './customServer';
 export * from './dataHandler';
 export * from './monitor';
+export * from './renderHandler';
